Agregue fetch en rutas relativas

+++ js/interacciones.js
@@ -16,46 +16,73 @@ let storage = JSON.parse(localStorage.getItem("Carrito"))
 
 storage ? (carrito = storage, mostrarCarrito()) : carrito = []
 
+
+
+document.addEventListener("click", (e) => {
+  if(e.target && e.target.matches("button#agregar")){
+      e.preventDefault();
+      carrito.push(producto);
+  localStorage.setItem("Carrito", JSON.stringify(carrito))
+  let subtotal = carrito.reduce((total, producto) => total + producto.precio, 0);
+
+
+  precioTotal.innerText = "Subtotal: $" + subtotal + "\nTotal: $" + suma(subtotal, sumaIva(subtotal));
+
+  Swal.fire({
+    icon: 'success',
+    title: 'Excelente',
+    text: 'Has añadido ' + producto.nombre + ' Al carrito\n' + ' te cuesta $' + producto.precio + '\nY con iva te sale en: $' + suma(producto.precio, sumaIva(producto.precio)),
+    confirmButtonText: 'Vale!',
+    confirmButtonColor: '#ff8836'
+  })
+  }
+  mostrarCarrito();
+  
+})
 // Hago la funcion para agregarlo al html
 function listado() {
-  listaPrecios.forEach((producto) => {
-    const catalogoProductos = document.createElement('div')
-    catalogoProductos.setAttribute("class", "card col-md-3 articulo")
-    catalogoProductos.innerHTML += ` 
-      <img src="${producto.img}">
-      <h3>${producto.nombre}</h3>
-      <h3>${producto.precio}</h3>
-      <button class="btn btn-dark " id=${producto.id}>Agregar al carrito</button>`
-    mostrarTodoProductos.appendChild(catalogoProductos)
+  fetch('../json/data.json')
+  .then( (res) => res.json())
+    .then( (data) => {data.forEach((producto) => {
+      let targeta = document.createElement("div");
+      mostrarTodoProductos.append(targeta);
+      targeta.setAttribute("class", "card col-md-3 articulo");
+      let img = document.createElement("img");
+      img.setAttribute("src", producto.img);
+      let nombre = document.createElement("h3");
+      nombre.innerText = producto.nombre;
+      let precio = document.createElement("h4");
+      precio.innerText = "$" + producto.precio;
+      let botonDeCompra = document.createElement("button");
+      botonDeCompra.innerText = "Agregar al carrito";
+      botonDeCompra.setAttribute("class", "btn btn-dark");
+      targeta.append(img, nombre, precio, botonDeCompra);
+  
+      botonDeCompra.onclick = () => {
+        carrito.push(producto);
+        localStorage.setItem("Carrito", JSON.stringify(carrito))
+        let subtotal = carrito.reduce(
+          (total, producto) => total + producto.precio,
+          0
+        );
+  
+  
+        precioTotal.innerText = "Subtotal: $" + subtotal + "\nTotal: $" + suma(subtotal, sumaIva(subtotal));
+  
+  
+        Swal.fire({
+          icon: 'success',
+          title: 'Excelente',
+          text: 'Has añadido ' + producto.nombre + ' Al carrito\n' + ' te cuesta $' + producto.precio + '\nY con iva te sale en: $' + suma(producto.precio, sumaIva(producto.precio)),
+          confirmButtonText: 'Vale!',
+          confirmButtonColor: '#ff8836'
+        })
+        mostrarCarrito();
+      };
+    });})
 
-    // let botonComprar= document.getElementById("agregar")
-    
-    document.addEventListener("click", (e) => {
-      if(e.target && e.target.matches("button#agregar")){
-          e.preventDefault();
-          carrito.push(producto);
-      localStorage.setItem("Carrito", JSON.stringify(carrito))
-      let subtotal = carrito.reduce((total, producto) => total + producto.precio, 0);
-
-
-      precioTotal.innerText = "Subtotal: $" + subtotal + "\nTotal: $" + suma(subtotal, sumaIva(subtotal));
-
-      Swal.fire({
-        icon: 'success',
-        title: 'Excelente',
-        text: 'Has añadido ' + producto.nombre + ' Al carrito\n' + ' te cuesta $' + producto.precio + '\nY con iva te sale en: $' + suma(producto.precio, sumaIva(producto.precio)),
-        confirmButtonText: 'Vale!',
-        confirmButtonColor: '#ff8836'
-      })
-      }
-      mostrarCarrito();
-      
-  })
-    
-      
-      
-
-  });
+  
+  
 }
 
 
@@ -73,8 +100,7 @@ function mostrarCarrito() {
     divCart.innerHTML += ` 
       <img src="${producto.img}">
       <h3>${producto.nombre}</h3>
-      <h3>$${producto.precio}</h3>
-      <button class="eliminar btn btn-dark" data-id=${producto.id}>X</button>`
+      <h3>$${producto.precio}</h3>`
 
 
     div.appendChild(divCart)
@@ -95,6 +121,3 @@ borrarCarrito.onclick = () => {
   precioTotal.innerText = "Total: $0";
 
 };
-
-
-
